refactor(dashboard): split route handlers and contracts
