+++ crates/agileplus-dashboard/src/routes/dashboard.rs
@@ -11,7 +11,6 @@
 //! partial template; otherwise return the full page layout.
 
 use std::collections::HashMap;
-use std::path::PathBuf;
 
 use axum::{
     extract::{Path, Query, State},
@@ -19,7 +18,6 @@ use axum::{
     response::{IntoResponse, Response},
 };
 use chrono::Utc;
-use serde::{Deserialize, Serialize};
 
 use crate::app_state::SharedState;
 use crate::templates::{
@@ -34,17 +32,6 @@ use super::helpers::{
 };
 
 // â”€â”€ JSON API Response Types â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
-
-/// JSON response for GET /api/dashboard/work-packages.json
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct WorkPackageJson {
-    pub id: String,
-    pub feature_id: i64,
-    pub title: String,
-    pub status: String,
-    pub priority: String,
-    pub assignee: Option<String>,
-}
 
 #[allow(dead_code)] // planned dashboard helper - will be wired in upcoming UI integration
 fn build_feature_events(
@@ -372,156 +359,3 @@ pub async fn time_footer() -> axum::response::Html<String> {
 }
 
 // â”€â”€ SSE Stream /api/stream â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
-
-use axum::response::sse::{Event, Sse};
-use std::convert::Infallible;
-use tokio::time::{Duration, interval};
-
-/// GET /api/stream (Server-Sent Events)
-/// Streams real-time feature and health updates to connected clients.
-/// Broadcasts heartbeat with feature count and health status every 5 seconds.
-pub async fn sse_stream(
-    State(state): State<SharedState>,
-) -> Sse<impl tokio_stream::Stream<Item = Result<Event, Infallible>>> {
-    let state = state.clone();
-    let stream = async_stream::stream! {
-        let mut ticker = interval(Duration::from_secs(5));
-        loop {
-            ticker.tick().await;
-            let store = state.read().await;
-
-            // Broadcast feature_updated event to refresh kanban
-            let feature_count = store.features.len();
-            let data = serde_json::json!({ "type": "heartbeat", "features": feature_count }).to_string();
-            yield Ok(Event::default()
-                .event("feature_updated")
-                .data(data));
-
-            // Broadcast health status
-            let healthy_count = store.health.iter().filter(|s| s.healthy).count();
-            let total_count = store.health.len();
-            let health_data = serde_json::json!({
-                "healthy": healthy_count,
-                "total": total_count,
-                "all_healthy": healthy_count == total_count
-            }).to_string();
-            yield Ok(Event::default()
-                .event("health_changed")
-                .data(health_data));
-        }
-    };
-    Sse::new(stream)
-}
-
-// â”€â”€ /api/dashboard/work-packages.json â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
-
-/// GET /api/dashboard/work-packages.json
-/// Returns all work packages across all features as a flat JSON array.
-/// Used by the React dashboard at port 5176 to populate the work-package store.
-pub async fn all_work_packages_json(State(state): State<SharedState>) -> impl IntoResponse {
-    let store = state.read().await;
-    let work_packages: Vec<WorkPackageJson> = store
-        .work_packages
-        .iter()
-        .flat_map(|(feature_id, wps)| {
-            wps.iter().map(|wp| {
-                let status = match wp.state {
-                    agileplus_domain::domain::work_package::WpState::Planned => "planned",
-                    agileplus_domain::domain::work_package::WpState::Doing => "in_progress",
-                    agileplus_domain::domain::work_package::WpState::Review => "in_progress",
-                    agileplus_domain::domain::work_package::WpState::Done => "completed",
-                    agileplus_domain::domain::work_package::WpState::Blocked => "blocked",
-                };
-                WorkPackageJson {
-                    id: wp.id.to_string(),
-                    feature_id: *feature_id,
-                    title: wp.title.clone(),
-                    status: status.to_string(),
-                    priority: "medium".to_string(),
-                    assignee: wp.agent_id.clone(),
-                }
-            })
-        })
-        .collect();
-
-    axum::Json(serde_json::json!({
-        "work_packages": work_packages,
-        "count": work_packages.len(),
-        "timestamp": chrono::Utc::now().to_rfc3339(),
-    }))
-}
-
-// â”€â”€ /api/dashboard/epics-stories.json â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
-
-/// GET /api/dashboard/epics-stories.json
-/// Reads Epics + Stories directly from the SQLite database and returns them
-/// as a flat JSON payload. Used by the React dashboard at port 5176.
-pub async fn epics_stories_json() -> impl IntoResponse {
-    // Resolve db path: DATABASE_URL env â†’ DATABASE_PATH env â†’ default agileplus.db
-    let db_path: PathBuf = if let Ok(url) = std::env::var("DATABASE_URL") {
-        url.strip_prefix("sqlite:").unwrap_or(&url).into()
-    } else if let Ok(p) = std::env::var("DATABASE_PATH") {
-        PathBuf::from(p)
-    } else {
-        PathBuf::from("agileplus.db")
-    };
-
-    let conn = match rusqlite::Connection::open(&db_path) {
-        Ok(c) => c,
-        Err(e) => {
-            return axum::Json(serde_json::json!({
-                "epics": [],
-                "stories": [],
-                "epic_count": 0,
-                "story_count": 0,
-                "error": format!("db open failed: {e}"),
-            }));
-        }
-    };
-
-    // Query epics
-    let epics: Vec<serde_json::Value> = {
-        let mut stmt = conn
-            .prepare("SELECT id, title, status, requirement_id FROM epics ORDER BY id")
-            .unwrap_or_else(|_| conn.prepare("SELECT 1 WHERE 0").unwrap());
-        stmt.query_map([], |row| {
-            Ok(serde_json::json!({
-                "id": row.get::<_, i64>(0).unwrap_or(0),
-                "title": row.get::<_, String>(1).unwrap_or_default(),
-                "status": row.get::<_, String>(2).unwrap_or_default(),
-                "requirement_id": row.get::<_, Option<String>>(3).unwrap_or(None),
-            }))
-        })
-        .map(|rows| rows.filter_map(|r| r.ok()).collect())
-        .unwrap_or_default()
-    };
-
-    // Query stories
-    let stories: Vec<serde_json::Value> = {
-        let mut stmt = conn
-            .prepare("SELECT id, epic_id, title, status, requirement_id FROM stories ORDER BY id")
-            .unwrap_or_else(|_| conn.prepare("SELECT 1 WHERE 0").unwrap());
-        stmt.query_map([], |row| {
-            Ok(serde_json::json!({
-                "id": row.get::<_, i64>(0).unwrap_or(0),
-                "epic_id": row.get::<_, Option<i64>>(1).unwrap_or(None),
-                "title": row.get::<_, String>(2).unwrap_or_default(),
-                "status": row.get::<_, String>(3).unwrap_or_default(),
-                "requirement_id": row.get::<_, Option<String>>(4).unwrap_or(None),
-            }))
-        })
-        .map(|rows| rows.filter_map(|r| r.ok()).collect())
-        .unwrap_or_default()
-    };
-
-    let epic_count = epics.len();
-    let story_count = stories.len();
-
-    axum::Json(serde_json::json!({
-        "epics": epics,
-        "stories": stories,
-        "epic_count": epic_count,
-        "story_count": story_count,
-        "timestamp": chrono::Utc::now().to_rfc3339(),
-    }))
-}
