+++ crates/agileplus-dashboard/src/routes/helpers.rs
@@ -271,75 +271,6 @@ pub fn percent_encode_path(path: &str) -> String {
         .collect()
 }
 
-// ── Plane configuration utilities ──────────────────────────────────────────
-
-pub(super) fn plane_api_key_hint(api_key: &Option<String>) -> String {
-    match api_key {
-        Some(key) => match (key.chars().next(), key.chars().next_back()) {
-            (Some(first), Some(last)) => format!("{first}••••••{last}"),
-            _ => "Configured".to_string(),
-        },
-        None => "Not configured".to_string(),
-    }
-}
-
-pub(super) fn plane_health_endpoints(
-    services: &[crate::app_state::ServiceHealth],
-) -> Vec<crate::templates::PlaneHealthEndpointView> {
-    services
-        .iter()
-        .filter(|service| service.name.contains("Plane") || service.name.starts_with("API"))
-        .map(|service| crate::templates::PlaneHealthEndpointView {
-            name: service.name.clone(),
-            healthy: service.healthy,
-            degraded: service.degraded,
-            latency_ms: service.latency_ms,
-            last_check_utc: service
-                .last_check
-                .format("%Y-%m-%d %H:%M:%S UTC")
-                .to_string(),
-        })
-        .collect()
-}
-
-pub(super) fn plane_sync_mode() -> String {
-    if parse_bool_env("PLANE_SYNC_BIDIRECTIONAL", false) {
-        "Bidirectional".to_string()
-    } else {
-        "One-way".to_string()
-    }
-}
-
-pub(super) fn plane_connection_checks(
-    api_key: &Option<String>,
-    workspace: &Option<String>,
-) -> (bool, String, Vec<String>) {
-    let mut warnings = Vec::new();
-    if api_key.is_none() {
-        warnings.push("Missing PLANE_API_KEY; configure a valid Plane API key".to_string());
-    }
-    if workspace.is_none() {
-        warnings.push("Missing PLANE_WORKSPACE; set workspace slug for Plane sync".to_string());
-    }
-
-    if warnings.is_empty() {
-        (true, "Connected via PLANE_API_KEY".to_string(), warnings)
-    } else if warnings.len() == 1 {
-        let status = warnings[0].clone();
-        (false, status, warnings)
-    } else {
-        (false, "Plane settings incomplete".to_string(), warnings)
-    }
-}
-
-pub(super) fn percentage_coverage(hit: usize, total: usize) -> String {
-    if total == 0 {
-        return "0/0 (0%)".to_string();
-    }
-    let ratio = (hit.saturating_mul(100)).saturating_div(total);
-    format!("{hit}/{total} ({ratio}%)")
-}
-
 // ── Service restart command validation ─────────────────────────────────────
 
 const ALLOWED_RESTART_PROGRAMS: [&str; 4] = ["systemctl", "docker", "process-compose", "echo"];
